fix: update relative imports after the selectors move to ui/selectors/money
The move one directory deeper left the shared-lib imports at the old depth.
